backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.2.16
langchain-openai==0.1.23
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0